# test/perf/perf_commitlog: Add a small commitlog throughput test