@avikivity 2022-03-20
Looks reasonable. Please add the result of a test (with insn count working, --smp 1) to the changelog to act as a baseline.